requirements.txt
pytest
pytest-selenium
pytest-variables
pytest-html
PyPOM